Code move and cleanup
Split `lib/dataplane/config/zone.h` into partitions
