+++ agent/balancer/main.go
@@ -3,6 +3,8 @@ package main
 //#cgo CFLAGS: -I../../ -I../../lib
 //#cgo LDFLAGS: -L../../build/modules/balancer/ -lbalancer_cp
 //#cgo LDFLAGS: -L../../build/lib/controlplane/agent -lagent
+//#cgo LDFLAGS: -L../../build/lib/controlplane/config -lconfig_cp
+//#cgo LDFLAGS: -L../../build/lib/dataplane/config -lconfig_dp
 //#include "api/agent.h"
 //#include "modules/balancer/controlplane.h"
 import "C"
