+++ lib/controlplane/config/meson.build
@@ -7,14 +7,14 @@ sources = files(
   'zone.c',
 )
 
-lib_config_dp = static_library(
-  'config_dp',
+lib_config_cp = static_library(
+  'config_cp',
   sources,
   dependencies: dependencies,
   install: false,
 )
 
-lib_config_dp_dep = declare_dependency(
-  link_with: lib_config_dp,
+lib_config_cp_dep = declare_dependency(
+  link_with: lib_config_cp,
   include_directories: yanet_libdir,
 )
