+++ lib/dataplane/config/topology.h
@@ -4,7 +4,4 @@
 
 struct dp_topology {
 	uint64_t device_count;
-
-	// Phy <-> Virt device mapping
-	uint16_t *forward_map;
 };
