+++ lib/dataplane/config/zone.c
@@ -1,0 +1,74 @@
+#include "zone.h"
+
+void
+dp_config_wait_for_gen(struct dp_config *dp_config, uint64_t gen) {
+	struct dp_worker **workers = ADDR_OF(&dp_config->workers);
+	uint64_t idx = 0;
+	do {
+		struct dp_worker *worker = ADDR_OF(workers + idx);
+		if (worker->gen < gen) {
+			// TODO cpu yield
+			continue;
+		}
+
+		++idx;
+	} while (idx < dp_config->worker_count);
+}
+
+bool
+dp_config_try_lock(struct dp_config *dp_config) {
+	pid_t pid = getpid();
+	pid_t zero = 0;
+	return __atomic_compare_exchange_n(
+		&dp_config->config_lock,
+		&zero,
+		pid,
+		false,
+		__ATOMIC_RELAXED,
+		__ATOMIC_RELAXED
+	);
+}
+
+void
+dp_config_lock(struct dp_config *dp_config) {
+	pid_t pid = getpid();
+	pid_t zero = 0;
+	while (!__atomic_compare_exchange_n(
+		&dp_config->config_lock,
+		&zero,
+		pid,
+		false,
+		__ATOMIC_RELAXED,
+		__ATOMIC_RELAXED
+	)) {
+		zero = 0;
+	};
+}
+
+bool
+dp_config_unlock(struct dp_config *dp_config) {
+	pid_t pid = getpid();
+	pid_t zero = 0;
+	return __atomic_compare_exchange_n(
+		&dp_config->config_lock,
+		&pid,
+		zero,
+		false,
+		__ATOMIC_RELAXED,
+		__ATOMIC_RELAXED
+	);
+}
+
+int
+dp_config_lookup_module(
+	struct dp_config *dp_config, const char *name, uint64_t *index
+) {
+	struct dp_module *modules = ADDR_OF(&dp_config->dp_modules);
+	for (uint64_t idx = 0; idx < dp_config->module_count; ++idx) {
+		if (!strncmp(modules[idx].name, name, 80)) {
+			*index = idx;
+			return 0;
+		}
+	}
+	return -1;
+}
